refactor: rename SimpleModel back to Model
